:zap: Improve isDotnetFile, fix #113
