+++ app/src/main/java/com/kyhsgeekcode/disassembler/APKFragment.kt
@@ -29,6 +29,7 @@ class APKFragment : Fragment() {
         super.onCreate(savedInstanceState)
         arguments?.let {
             relPath = it.getString(ARG_PARAM1)!!
+            it.clear()
         }
     }
