+++ app/src/main/java/com/kyhsgeekcode/disassembler/BinaryDisasmFragment.kt
@@ -51,6 +51,7 @@ class BinaryDisasmFragment : Fragment(), IOnBackPressed {
         arguments?.let {
             relPath = it.getString(ARG_PARAM)!!
             parsedFile = (parentFragment as IParsedFileProvider).parsedFile
+            it.clear()
         }
     }
