+++ app/src/main/java/com/kyhsgeekcode/disassembler/BinaryFragment.kt
@@ -26,6 +26,7 @@ class BinaryFragment : Fragment(), ITabController, IParsedFileProvider {
         super.onCreate(savedInstanceState)
         arguments?.let {
             relPath = it.getString(ARG_PARAM1)!!
+            it.clear()
         }
     }
