+++ app/src/main/java/com/kyhsgeekcode/disassembler/DotNetFragment.kt
@@ -29,6 +29,7 @@ class DotNetFragment : Fragment() {
         super.onCreate(savedInstanceState)
         arguments?.let {
             relPath = it.getString(ARG_PARAM1)!!
+            it.clear()
         }
     }
