+++ app/src/main/java/com/kyhsgeekcode/disassembler/TextFragment.kt
@@ -35,6 +35,7 @@ class TextFragment : Fragment() {
         super.onCreate(savedInstanceState)
         arguments?.let {
             relPath = it.getString(ARG_PARAM)!!
+            it.clear()
         }
         Log.d(TAG, "relPath:$relPath")
         fileContent = ProjectDataStorage.getFileContent(relPath)
